requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.8
contourpy==1.0.7
cycler==0.11.0
deep-translator==1.11.4
docstring_parser==0.16
dotenv==0.9.9
filelock==3.18.0
Flask==3.1.0
fonttools==4.39.3
fsspec==2025.3.2
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
huggingface-hub==0.30.2
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.4
MarkupSafe==3.0.2
matplotlib==3.7.1
mpmath==1.3.0
nltk==3.9.1
numpy==1.24.2
packaging==24.2
pandas==2.2.3
Pillow==9.5.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.2
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.10.1
seaborn==0.13.2
shapely==2.1.0
six==1.16.0
soupsieve==2.6
sympy==1.11.1
threadpoolctl==3.6.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.15
vertexai==1.71.1
Werkzeug==3.1.3